improvements to command lib
fixed minor bugs in ArgumentParser

changed flloating params prefixed with "-", flags prefixed with "--"; floating params use more space so it makes sense to use the smaller prefix

removed CommandArgument#getDefaultValue and replaced with CommandArgument#isDefaultValue;

removed CommandArguments#hasString and #hasValue as redundant; always returns true after changes to the way default values are handled. Use CommandArguments#isDefaultValue to determine if the user supplied a value.

internal changes to ArgumentParser; argument map always contains a value for a parameter, even if it was optional and omitted. No need to check if value is null and substitute default value anymore. Default value is inserted by the parser.

fix exception not thrown when required static param is omitted and no arguments provided.

throw RuntimeException if a parameter name is requested from CommandArguments that is not defined by the command; helps with debugging.

CommandInfoContainer can have CommandInfo directly passed into a constructor, for test purposes.

add UsageGeneratorTest, CommandsArgumentTest and ArgumentParserTest as well as more dummy classes and utils to support testing

 code cleanup